FEAT: Make spot event to be movable up
